backend/rag-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
scikit-learn==1.5.2
numpy==2.1.3
pypdf==5.1.0
python-docx==1.1.2
